refactor(all): new structure
