+++ src/index.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import './styles/index.css';
 import App from './App';
-import Registration from './registration.js';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
